fix: Resolve compiler errors and mark methods as static
- Fix BeliefState missing DetectedEntities/PrimaryTarget in CueRegistry.cs
  by using ThreatLevel instead
- Fix Detection.Distance -> EstimatedDistance in ReferenceBuilder.cs
- Fix ambiguous 'utils' reference in PPOTrainer.cs using nn.utils
- Fix double to float conversion in InteroceptiveProcessor.cs
- Convert ActionPlan and BeliefState from class to record for 'with' expressions
- Add Microsoft.Extensions.Logging.Abstractions package reference
- Mark 20+ methods as static (CA1822) across multiple files for better performance

+++ src/OKET.Agent/Cognition/InteroceptiveProcessor.cs
@@ -206,8 +206,8 @@ public sealed class InteroceptiveProcessor
         var entries = _history.ToList();
         int midpoint = entries.Count / 2;
 
-        float recentHealth = entries.Skip(midpoint).Average(h => h.Health);
-        float olderHealth = entries.Take(midpoint).Average(h => h.Health);
+        float recentHealth = (float)entries.Skip(midpoint).Average(h => h.Health);
+        float olderHealth = (float)entries.Take(midpoint).Average(h => h.Health);
 
         float healthTrend = (recentHealth - olderHealth) / 50f;
